requirements.txt
# pip installable libraries required in this project:

Pillow>=10.0.0
prettytable>=3.0.0
